requirements.txt
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.4
opencv-python>=4.8.0
pillow>=10.0.0
Flask==2.3.3
Werkzeug==2.3.7
torch>=2.2.0
torchvision>=0.17.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
matplotlib>=3.7.2
python-dotenv==1.0.0